Refactor Dashboard Component and Implement User Authentication Views
- Replaced NormalWidthDashboard and SmallerWidthDashboard with RegisteredDashboard and NotRegisteredDashboard components for better user experience based on authentication status.
- Introduced a LeaderboardSection component to display leaderboard data for both logged-in and unlogged users.
- Removed unused demo pages and components to streamline the dashboard structure.
- Enhanced user data fetching logic to handle authenticated and unauthenticated states more effectively.
